+++ src/app/shared/module/poe/service/vendor-recipe/processors/chance-recipe-processor.service.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@angular/core'
-import { BaseItemTypesService } from '@shared/module/poe/service/base-item-types/base-item-types.service'
 import { LoggerService } from '@app/service'
+import { BaseItemTypesService } from '@shared/module/poe/service/base-item-types/base-item-types.service'
+import { ClientStringService } from '@shared/module/poe/service/client-string/client-string.service'
+import { ContextService } from '@shared/module/poe/service/context.service'
 import { ItemLevelBasedRecipeProcessorService } from './item-level-based-recipe-processor.service'
 
 @Injectable({
@@ -9,9 +11,11 @@ import { ItemLevelBasedRecipeProcessorService } from './item-level-based-recipe-
 export class ChanceRecipeProcessorService extends ItemLevelBasedRecipeProcessorService {
   constructor(
     readonly baseItemTypeService: BaseItemTypesService,
+    readonly clientString: ClientStringService,
+    readonly context: ContextService,
     readonly logger: LoggerService
   ) {
-    super(baseItemTypeService, logger)
+    super(baseItemTypeService, clientString, context, logger)
   }
 
   protected get minItemLevel(): number {