+++ src/app/shared/module/poe/service/vendor-recipe/processors/exalted-shard-recipe-processor.service.ts
@@ -1,9 +1,11 @@
-import { Injectable } from '@angular/core';
-import { LoggerService } from '@app/service';
-import { BaseItemTypesService } from '@shared/module/poe/service/base-item-types/base-item-types.service';
-import { ItemSetRecipeUserSettings, RecipeItemGroup } from '@shared/module/poe/type';
-import { ItemSetRecipeProcessorService } from './item-set-recipe-processor.service';
-import { ExpandedStashItem } from './recipe-processor.service';
+import { Injectable } from '@angular/core'
+import { LoggerService } from '@app/service'
+import { BaseItemTypesService } from '@shared/module/poe/service/base-item-types/base-item-types.service'
+import { ClientStringService } from '@shared/module/poe/service/client-string/client-string.service'
+import { ContextService } from '@shared/module/poe/service/context.service'
+import { ItemSetRecipeUserSettings, RecipeItemGroup } from '@shared/module/poe/type'
+import { ItemSetRecipeProcessorService } from './item-set-recipe-processor.service'
+import { ExpandedStashItem } from './recipe-processor.service'
 
 @Injectable({
   providedIn: 'root',
@@ -11,9 +13,11 @@ import { ExpandedStashItem } from './recipe-processor.service';
 export class ExaltedShardRecipeProcessorService extends ItemSetRecipeProcessorService {
   constructor(
     readonly baseItemTypeService: BaseItemTypesService,
+    readonly clientString: ClientStringService,
+    readonly context: ContextService,
     readonly logger: LoggerService,
   ) {
-    super(baseItemTypeService, logger)
+    super(baseItemTypeService, clientString, context, logger)
   }
 
   protected getSplittedCandidates(allCandidates: ExpandedStashItem[], settings: ItemSetRecipeUserSettings): ExpandedStashItem[][] {